requirements_lock.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --allow-unsafe --no-annotate --output-file=requirements_lock.txt requirements.txt
#
greenlet==1.1.2
psycopg2-binary==2.9.3
sqlalchemy[postgresql_psycopg2binary]==1.4.36
